requirements.txt
Flask==0.8
Jinja2==2.6
PyYAML==3.10
Werkzeug==0.8.1
distribute==0.6.19
wsgiref==0.1.2
